Split gait generator plugin to a view and a controller
